--- src/main/java/Command.java
@@ -1,8 +1,0 @@
-public abstract class Command {
-
-    abstract void execute(TaskList tasks, Ui ui, Storage storage);
-    public boolean isExit() {
-        return false;
-    }
-
-}
