--- src/main/java/ExitCommand.java
@@ -1,10 +1,0 @@
-public class ExitCommand extends Command{
-
-    void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.respond("Bye. Hope to see you again soon!");
-    }
-
-    public boolean isExit() {
-        return true;
-    }
-}
