--- src/main/java/ListCommand.java
@@ -1,7 +1,0 @@
-public class ListCommand extends Command{
-
-    @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.respond(tasks.listing());
-    }
-}
