--- src/main/java/PrintCommand.java
@@ -1,11 +1,0 @@
-public class PrintCommand extends Command{
-    private final String text;
-
-    PrintCommand(String s) {
-        text = s;
-    }
-
-    void execute(TaskList tasks, Ui ui, Storage storage) {
-        ui.respond(text);
-    }
-}
