+++ src/main/java/commands/AddCommand.java
@@ -1,13 +1,20 @@
-public class AddCommand extends Command{
+package commands;
+
+import data.Task;
+import data.TaskList;
+import storage.Storage;
+import ui.Ui;
+
+public class AddCommand extends Command {
     private final Task t;
 
-    AddCommand(Task task) {
+    public AddCommand(Task task) {
         super();
         this.t = task;
     }
 
     @Override
-    void execute(TaskList tasks, Ui ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         tasks.add(t);
         storage.record(t + "\n");
         ui.respond("Got it. I've added this task:\n  " +
