+++ src/main/java/commands/Command.java
@@ -1,0 +1,14 @@
+package commands;
+
+import data.TaskList;
+import storage.Storage;
+import ui.Ui;
+
+public abstract class Command {
+
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage);
+    public boolean isExit() {
+        return false;
+    }
+
+}
