+++ src/main/java/commands/ListCommand.java
@@ -1,0 +1,13 @@
+package commands;
+
+import data.TaskList;
+import storage.Storage;
+import ui.Ui;
+
+public class ListCommand extends Command {
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.respond(tasks.listing());
+    }
+}
