+++ src/main/java/common/NoGoodException.java
@@ -1,3 +1,5 @@
+package common;
+
 public class NoGoodException extends Exception {
     public NoGoodException(String errorMessage) {
         super(errorMessage);
