+++ src/main/java/common/Type.java
@@ -1,3 +1,5 @@
+package common;
+
 public enum Type {
     TODO,
     DEADLINE,
